Major refac on the way
Major refactoring on the way.
Scene objects can now be transformed into hittable pointers.
I need to figure out transforming SceneObjects to a holder of
device pointers.
The device pointers should hold loaded device pointers.

+++ include/nextweek/box.cuh
@@ -45,7 +45,7 @@ public:
     }
   }
   __host__ __device__ void rotate_y(Hittable **&ss,
-                                    double angle) {
+                                    float angle) {
     for (int i = start_index; i < end_index; i++) {
       list[i] = new RotateY(list[i], angle);
       ss[i] = list[i];
